Usability improvements (see details)
This commit adds default settings that are shared via a config.sh file,
this allows for clearer output messages from the newimage/liveosimage
scripts as well as providing the flexibility to tweak the host OS setup.

A default config.sh file has been provided.

Additionally, DNS/HTTP forwarding has been added to allow for clients to
reach the internet (should clients need to).

Relevant documentation and examples are provided in the README.